osce-video-grader/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.11
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
ctranslate2==4.4.0
cycler==0.12.1
dataclasses==0.8
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
durationpy==0.10
einops==0.8.1
executing==2.2.0
fastapi==0.115.9
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.25.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
hyperframe==6.1.0
hyperpyyaml==1.2.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
kubernetes==32.0.1
lazy-loader==0.4
levenshtein==0.27.1
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.15
mmh3==5.1.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
optuna==4.3.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
pooch==1.8.2
portalocker==2.10.1
posthog==4.2.0
primepy==1.3
proglog==0.1.12
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyannote-audio==3.3.2
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyparsing==3.2.3
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-levenshtein==0.27.1
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
pyvips==3.0.0
pyyaml==6.0.2
pyzmq==26.4.0
qdrant-client==1.14.2
rapidfuzz==3.13.0
ratelimit==2.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel-yaml==0.18.11
ruamel-yaml-clib==0.2.12
safetensors==0.5.3
scenedetect==0.6.6
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
sqlalchemy==2.0.41
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.45.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardx==2.6.2.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
whisperx==3.3.4
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
